pyproject.toml
[project]
name = "agenthazard"
version = "0.1.0"
description = "Source code for Paper: Mobile GUI Agents under Real-world Threats: Are We There Yet?"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "click>=8.3.1",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "structlog>=25.5.0",
    "tabulate>=0.10.0",
]
authors = [
    { name = "Zsbyqx20", email = "112002598+Zsbyqx20@users.noreply.github.com" },
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ah = "agenthazard.cli:main"

[tool.pyright]
include = ["src/agenthazard"]

[tool.poe.tasks]
_ruff_check_format = "ruff format . --check --preview"
_ruff_check = "ruff check . --preview"
_pyright = "pyright"
_ruff_sort_imports = "ruff check --select I --fix . --preview"
_ruff_format_code = "ruff format . --preview"
fix = "ruff check --fix . --preview"
fix_unsafe = "ruff check --preview --fix --unsafe-fixes ."
export = "uv pip compile pyproject.toml -o requirements.txt"

[[tool.poe.tasks.format]]
sequence = ['_ruff_sort_imports', '_ruff_format_code']
ignore_fail = 'return_non_zero'

[[tool.poe.tasks.check]]
sequence = ['_ruff_check_format', '_ruff_check', '_pyright']
ignore_fail = 'return_non_zero'

[dependency-groups]
dev = [
    "poethepoet>=0.42.1",
    "pyright>=1.1.408",
    "ruff>=0.15.6",
]
